Extract native search command handling
