+++ crates/truthdb-core/src/engine/search/command.rs
@@ -1,0 +1,256 @@
+use std::collections::BTreeMap;
+
+use serde_json::Value;
+use thiserror::Error;
+
+use super::super::EngineError;
+use super::index::{Document, FieldType, SearchQuery, value_type_name};
+
+#[derive(Debug, Clone)]
+pub(in crate::engine) enum Command {
+    CreateIndex {
+        name: String,
+        mappings: BTreeMap<String, FieldType>,
+    },
+    InsertDocument {
+        index: String,
+        document: Document,
+    },
+    Search {
+        index: String,
+        query: SearchQuery,
+    },
+}
+
+#[derive(Debug, Error)]
+pub enum CommandError {
+    #[error("invalid command: {0}")]
+    InvalidCommand(String),
+
+    #[error("malformed json: {0}")]
+    MalformedJson(String),
+
+    #[error("index '{0}' already exists")]
+    IndexAlreadyExists(String),
+
+    #[error("unknown index '{0}'")]
+    UnknownIndex(String),
+
+    #[error("unknown field '{0}'")]
+    UnknownField(String),
+
+    #[error("invalid field type for '{field}': expected {expected}, got {actual}")]
+    InvalidFieldType {
+        field: String,
+        expected: String,
+        actual: String,
+    },
+}
+
+/// Parses a legacy ES command. Returns `Ok(None)` when the input is not an
+/// ES command (a `{`-bodied create index / insert document / search) — the
+/// caller then routes it to the SQL engine. `Ok(Some(_))` is a well-formed
+/// ES command; `Err` is a malformed one.
+pub(in crate::engine) fn parse_command(input: &str) -> Result<Option<Command>, CommandError> {
+    let trimmed = input.trim();
+    if trimmed.is_empty() {
+        return Err(CommandError::InvalidCommand(
+            "command cannot be empty".to_string(),
+        ));
+    }
+
+    if let Some((header, body)) = split_command(trimmed, "create index")? {
+        let name = parse_single_name(header, "create index")?;
+        let mappings = parse_create_index_body(body)?;
+        return Ok(Some(Command::CreateIndex { name, mappings }));
+    }
+
+    if let Some((header, body)) = split_command(trimmed, "insert document")? {
+        let index = parse_single_name(header, "insert document")?;
+        let document = parse_document_body(body)?;
+        return Ok(Some(Command::InsertDocument { index, document }));
+    }
+
+    if let Some((header, body)) = split_command(trimmed, "search")? {
+        let index = parse_single_name(header, "search")?;
+        let query = parse_search_body(body)?;
+        return Ok(Some(Command::Search { index, query }));
+    }
+
+    // Not an ES command: route to SQL.
+    Ok(None)
+}
+
+fn split_command<'a>(
+    input: &'a str,
+    prefix: &str,
+) -> Result<Option<(&'a str, &'a str)>, CommandError> {
+    let Some(body_start) = input.find('{') else {
+        return Ok(None);
+    };
+    let (header, body) = input.split_at(body_start);
+    let header = header.trim();
+    if !header
+        .to_ascii_lowercase()
+        .starts_with(&prefix.to_ascii_lowercase())
+    {
+        return Ok(None);
+    }
+    if body.trim().is_empty() {
+        return Err(CommandError::InvalidCommand(format!(
+            "{prefix} command requires a json body"
+        )));
+    }
+    Ok(Some((header, body)))
+}
+
+fn parse_single_name(header: &str, prefix: &str) -> Result<String, CommandError> {
+    let rest = header[prefix.len()..].trim();
+    if rest.is_empty() {
+        return Err(CommandError::InvalidCommand(format!(
+            "{prefix} command requires a name"
+        )));
+    }
+    if rest.split_whitespace().count() != 1 {
+        return Err(CommandError::InvalidCommand(format!(
+            "{prefix} command takes exactly one name before the json body"
+        )));
+    }
+    Ok(rest.to_string())
+}
+
+fn parse_create_index_body(body: &str) -> Result<BTreeMap<String, FieldType>, CommandError> {
+    let value = parse_json(body)?;
+    let root = as_object(&value, "create index body")?;
+    let mappings = root
+        .get("mappings")
+        .ok_or_else(|| CommandError::InvalidCommand("missing mappings object".to_string()))?;
+    let mappings = as_object(mappings, "mappings")?;
+    let properties = mappings.get("properties").ok_or_else(|| {
+        CommandError::InvalidCommand("missing mappings.properties object".to_string())
+    })?;
+    let properties = as_object(properties, "mappings.properties")?;
+
+    let mut fields = BTreeMap::new();
+    for (field_name, field_value) in properties {
+        let field_obj = as_object(field_value, &format!("field mapping '{field_name}'"))?;
+        let Some(field_type_value) = field_obj.get("type") else {
+            return Err(CommandError::InvalidCommand(format!(
+                "field mapping '{field_name}' is missing type"
+            )));
+        };
+        let Some(field_type_str) = field_type_value.as_str() else {
+            return Err(CommandError::InvalidCommand(format!(
+                "field mapping '{field_name}' type must be a string"
+            )));
+        };
+        let field_type = parse_field_type(field_type_str)?;
+        fields.insert(field_name.clone(), field_type);
+    }
+
+    Ok(fields)
+}
+
+fn parse_document_body(body: &str) -> Result<Document, CommandError> {
+    let value = parse_json(body)?;
+    let object = as_object(&value, "document body")?;
+    Ok(object.clone())
+}
+
+fn parse_search_body(body: &str) -> Result<SearchQuery, CommandError> {
+    let value = parse_json(body)?;
+    let root = as_object(&value, "search body")?;
+    let query = root
+        .get("query")
+        .ok_or_else(|| CommandError::InvalidCommand("missing query object".to_string()))?;
+    parse_search_query(query)
+}
+
+fn parse_search_query(value: &Value) -> Result<SearchQuery, CommandError> {
+    let object = as_object(value, "query")?;
+
+    if let Some(match_value) = object.get("match") {
+        let field_map = as_object(match_value, "match")?;
+        if field_map.len() != 1 {
+            return Err(CommandError::InvalidCommand(
+                "match query must contain exactly one field".to_string(),
+            ));
+        }
+        let (field, query_value) = field_map.iter().next().unwrap();
+        let Some(query) = query_value.as_str() else {
+            return Err(CommandError::InvalidCommand(format!(
+                "match query for field '{field}' must be a string"
+            )));
+        };
+        return Ok(SearchQuery::Match {
+            field: field.clone(),
+            query: query.to_string(),
+        });
+    }
+
+    if let Some(term_value) = object.get("term") {
+        let field_map = as_object(term_value, "term")?;
+        if field_map.len() != 1 {
+            return Err(CommandError::InvalidCommand(
+                "term query must contain exactly one field".to_string(),
+            ));
+        }
+        let (field, value) = field_map.iter().next().unwrap();
+        return Ok(SearchQuery::Term {
+            field: field.clone(),
+            value: value.clone(),
+        });
+    }
+
+    if let Some(bool_value) = object.get("bool") {
+        let bool_map = as_object(bool_value, "bool")?;
+        let must = parse_query_array(bool_map.get("must"), "bool.must")?;
+        let filter = parse_query_array(bool_map.get("filter"), "bool.filter")?;
+        return Ok(SearchQuery::Bool { must, filter });
+    }
+
+    Err(CommandError::InvalidCommand(
+        "query must contain one of: match, term, bool".to_string(),
+    ))
+}
+
+fn parse_query_array(value: Option<&Value>, label: &str) -> Result<Vec<SearchQuery>, CommandError> {
+    let Some(value) = value else {
+        return Ok(Vec::new());
+    };
+    let Some(items) = value.as_array() else {
+        return Err(CommandError::InvalidCommand(format!(
+            "{label} must be an array"
+        )));
+    };
+    items.iter().map(parse_search_query).collect()
+}
+
+fn parse_json(body: &str) -> Result<Value, CommandError> {
+    serde_json::from_str(body).map_err(|err| CommandError::MalformedJson(err.to_string()))
+}
+
+fn parse_field_type(raw: &str) -> Result<FieldType, CommandError> {
+    match raw {
+        "text" => Ok(FieldType::Text),
+        "keyword" => Ok(FieldType::Keyword),
+        "float" => Ok(FieldType::Float),
+        other => Err(CommandError::InvalidCommand(format!(
+            "unsupported field type '{other}'"
+        ))),
+    }
+}
+
+fn as_object<'a>(value: &'a Value, label: &str) -> Result<&'a Document, CommandError> {
+    value.as_object().ok_or_else(|| {
+        CommandError::InvalidCommand(format!(
+            "{label} must be a json object, got {}",
+            value_type_name(value)
+        ))
+    })
+}
+
+pub(in crate::engine) fn render_json(value: &Value) -> Result<String, EngineError> {
+    serde_json::to_string_pretty(value)
+        .map_err(|err| EngineError::Replay(format!("failed to render json response: {err}")))
+}
