+++ crates/truthdb-core/src/engine/search/mod.rs
@@ -1,7 +1,8 @@
 //! Native document-search implementation owned by [`super::Engine`].
 
+mod command;
 mod index;
 
-pub(super) use index::{
-    Document, EngineState, FieldType, IndexState, SearchQuery, value_type_name,
-};
+pub use command::CommandError;
+pub(super) use command::{Command, parse_command, render_json};
+pub(super) use index::{Document, EngineState, FieldType, IndexState, SearchQuery};
